refactor: split manuscript resolvers into separate files
Mock pubsub functionality to avoid issue with tearing down the connection while an IPC message delivery is outstanding
