+++ package.json
@@ -77,7 +77,7 @@
   ],
   "dependencies": {
     "@elifesciences/elife-theme": "^1.0.0",
-    "@pubsweet/ui": "^8.7.1",
+    "@pubsweet/ui": "^8.8.0",
     "@pubsweet/ui-toolkit": "^1.2.0",
     "apollo-link-state": "^0.4.1",
     "config": "^2.0.1",
@@ -92,8 +92,8 @@
     "newrelic": "^4.8.1",
     "omit-deep-lodash": "^1.1.2",
     "prop-types": "^15.6.2",
-    "pubsweet-client": "^6.0.1",
-    "pubsweet-server": "^10.0.1",
+    "pubsweet-client": "^6.1.0",
+    "pubsweet-server": "^10.0.3",
     "pug": "^2.0.3",
     "react": "^16.4.2",
     "react-apollo": "^2.1.11",
@@ -152,7 +152,7 @@
     "lint-staged": "^7.2.2",
     "mime": "^2.3.1",
     "prettier": "^1.8.2",
-    "pubsweet": "^3.0.6",
+    "pubsweet": "^3.0.8",
     "react-styleguidist": "^7.1.0",
     "replay": "^2.1.4",
     "string-replace-loader": "^2.1.1",
