+++ server/meca/package.json
@@ -18,7 +18,7 @@
     "xmlbuilder": "^10.0.0"
   },
   "devDependencies": {
-    "@pubsweet/db-manager": "^2.0.6",
+    "@pubsweet/db-manager": "^2.0.8",
     "body-parser": "^1.18.3",
     "express": "^4.16.3",
     "memory-fs": "^0.4.1",
