+++ server/xpub-server/index.js
@@ -1,11 +1,6 @@
 const { merge } = require('lodash')
 const fs = require('fs')
 
-const filterFileName = filename => entity =>
-  fs.existsSync(`${__dirname}/entities/${entity}/${filename}`)
-
-const entities = ['file', 'identity', 'manuscript', 'team', 'user']
-
 // concatenate schemas
 const xpubTypeDefs = fs.readFileSync(
   `${__dirname}/schema/xpub.graphqls`,
@@ -15,12 +10,10 @@ const elifeTypeDefs = fs.readFileSync(
   `${__dirname}/schema/elife.graphqls`,
   'utf8',
 )
-
-const entityTypeDefs = entities
-  .filter(filterFileName('typeDefs.graphqls'))
-  .map(name =>
-    fs.readFileSync(`${__dirname}/entities/${name}/typeDefs.graphqls`, 'utf8'),
-  )
+const entityTypeDefs = [
+  fs.readFileSync(`${__dirname}/entities/manuscript/typeDefs.graphqls`, 'utf8'),
+  fs.readFileSync(`${__dirname}/entities/user/typeDefs.graphqls`, 'utf8'),
+]
 const typeDefs = `
   ${xpubTypeDefs} 
   ${elifeTypeDefs} 
@@ -30,16 +23,12 @@ const typeDefs = `
 // merge resolvers
 const resolvers = merge(
   {},
-  ...entities
-    .filter(filterFileName('resolvers.js'))
-    .map(name => require(`./entities/${name}/resolvers.js`)),
+  require(`./entities/manuscript/resolvers`),
+  require(`./entities/user/resolvers`),
 )
 
 const registerRoutes = app => {
-  entities
-    .filter(filterFileName('routes.js'))
-    .forEach(name => require(`./entities/${name}/routes.js`)(app))
-
+  require(`./entities/user/routes`)(app)
   require('./routes')(app)
 }
